+++ src/ska_mid_dish_manager/devices/DishManagerDS.py
@@ -6,6 +6,7 @@ and the subservient devices
 
 import json
 import weakref
+from datetime import datetime
 from functools import reduce
 from typing import List, Optional, Tuple
 
@@ -46,7 +47,11 @@ from ska_mid_dish_manager.models.dish_enums import (
 )
 from ska_mid_dish_manager.release import ReleaseInfo
 from ska_mid_dish_manager.utils.command_logger import BaseInfoIt
-from ska_mid_dish_manager.utils.decorators import record_mode_change_request
+from ska_mid_dish_manager.utils.decorators import (
+    record_mode_change_request,
+    requires_component_manager,
+)
+from ska_mid_dish_manager.utils.schedulers import WatchdogTimerInactiveError
 from ska_mid_dish_manager.utils.track_table_input_validation import (
     TrackLoadTableFormatting,
     TrackTableTimestampError,
@@ -372,6 +377,8 @@ class DishManager(SKAController):
                 "dscpowerlimitkw": "dscPowerLimitKw",
                 "tracktablecurrentindex": "trackTableCurrentIndex",
                 "tracktableendindex": "trackTableEndIndex",
+                "lastwatchdogreset": "lastWatchdogReset",
+                "watchdogtimeout": "watchdogTimeout",
             }
             for attr in device._component_state_attr_map.values():
                 device.set_change_event(attr, True, False)
@@ -1380,6 +1387,39 @@ class DishManager(SKAController):
                 f" valid range is [{DSC_MIN_POWER_LIMIT_KW}, {DSC_MAX_POWER_LIMIT_KW}]."
             )
 
+    @attribute(
+        dtype=float,
+        access=AttrWriteType.READ,
+        doc="Returns the timestamp of the last watchdog reset in unix epoch seconds.",
+    )
+    def lastWatchdogReset(self):
+        """Returns lastWatchdogReset."""
+        return self.component_manager.component_state["lastwatchdogreset"]
+
+    @attribute(
+        dtype=float,
+        access=AttrWriteType.READ_WRITE,
+        doc="Dish manager watchdog timeout interval in seconds. "
+        "By writing a value greater than 0, the watchdog will be enabled. If the watchdog "
+        "is not reset within this interval, the dish will Stow on expiry of the timer. "
+        "The watchdog timer can be reset by calling the ResetWatchdog command. The watchdog "
+        "is disabled by writing a value less than or equal to 0.",
+    )
+    @requires_component_manager
+    def watchdogTimeout(self):
+        """Returns watchdogTimeout."""
+        return self.component_manager.component_state["watchdogtimeout"]
+
+    @watchdogTimeout.write
+    @requires_component_manager
+    def watchdogTimeout(self, value):
+        """Writes watchdogTimeout."""
+        self.component_manager._update_component_state(watchdogtimeout=value)
+        if value <= 0:
+            self.component_manager.watchdog_timer.disable()
+        else:
+            self.component_manager.watchdog_timer.enable(value)
+
     # --------
     # Commands
     # --------
@@ -1903,6 +1943,30 @@ class DishManager(SKAController):
         """
         if hasattr(self, "component_manager"):
             self.component_manager.sync_component_states()
+
+    @command(
+        dtype_out="DevVarLongStringArray",
+        display_level=DispLevel.OPERATOR,
+        doc_in="This command resets the watchdog timer. "
+        "lastWatchdogReset attribute will be updated with the unix timestamp."
+        "By default the watchdog timer is disabled and can be enabled by setting the "
+        "watchdogTimeout attribute to a value greater than 0.",
+        doc_out="Returns a DevVarLongStringArray with the return code and message.",
+    )
+    @BaseInfoIt(show_args=True, show_kwargs=True, show_ret=True)
+    @requires_component_manager
+    def ResetWatchdogTimer(self) -> DevVarLongStringArrayType:
+        """This command resets the watchdog timer."""
+        value = datetime.now().timestamp()
+        try:
+            self.component_manager.watchdog_timer.reset()
+        except WatchdogTimerInactiveError:
+            return ([ResultCode.FAILED], ["Watchdog timer is not active."])
+        self.component_manager._update_component_state(lastwatchdogreset=value)
+        return (
+            [ResultCode.OK],
+            [f"Watchdog timer reset at {value}s"],
+        )
 
     @command(dtype_out="DevVarLongStringArray")
     @BaseInfoIt(show_args=True, show_kwargs=True, show_ret=True)