+++ src/ska_mid_dish_manager/utils/schedulers.py
@@ -3,23 +3,32 @@
 import threading
 from typing import Callable
 
+DEFAULT_WATCHDOG_TIMEOUT = 10.0  # seconds
+
+
+class WatchdogTimerInactiveError(RuntimeError):
+    """Exception raised when the watchdog timer is not enabled."""
+
 
 class WatchdogTimer:
-    def __init__(self, timeout: float, callback_on_timeout: Callable):
+    def __init__(
+        self, callback_on_timeout: Callable = None, timeout: float = DEFAULT_WATCHDOG_TIMEOUT
+    ):
         """This class implements a watchdog timer that will make a callback
         when the timer expires.
 
-        :param timeout: The default time in seconds after which the callback will be
-        invoked if not reset.
-        :type timeout: float
         :param callback_on_timeout: The callback function to be invoked when the timer expires.
         :type callback_on_timeout: Callable
+        :param timeout: The default time in seconds after which the callback will be
+        invoked if not reset, defaults to DEFAULT_WATCHDOG_TIMEOUT
+        :type timeout: float, optional
+        :raises ValueError: If timeout specified is less than zero.
         """
         if timeout <= 0:
             raise ValueError("Timeout must be greater than 0.")
 
         self._timeout = timeout
-        self.callback_on_timeout = callback_on_timeout
+        self._callback_on_timeout = callback_on_timeout
         self._timer = None
         self._lock = threading.RLock()
         self._enabled = False
@@ -53,12 +62,12 @@ class WatchdogTimer:
     def reset(self):
         """Reset the watchdog timer.
 
-        :raises RuntimeError: If called when the timer is not enabled.
+        :raises WatchdogTimerInactive: If called when the timer is disabled.
         """
         with self._lock:
             if not self._enabled:
-                raise RuntimeError("Watchdog timer is not enabled. Call enable first.")
+                raise WatchdogTimerInactiveError("Watchdog timer is disabled. Call enable first.")
             if self._timer is not None:
                 self._timer.cancel()
-            self._timer = threading.Timer(self._timeout, self.callback_on_timeout)
+            self._timer = threading.Timer(self._timeout, self._callback_on_timeout)
             self._timer.start()